DatabasePerformanceAnalysis/requirements.txt
azure.identity
pyodbc
matplotlib
pandas